pref: optimize the ui in moment editing mode
